Misc. other cleanup
