+++ src/queries/responses.hpp
@@ -15,6 +15,14 @@ protected:
     nlohmann::json json_;
 };
 
+struct OpenAIError: public ErrorResponse {
+    OpenAIError(const std::string &response, const int status_code);
+};
+
+struct OllamaError: public ErrorResponse {
+    OllamaError(const std::string &response, const int status_code);
+};
+
 class SuccessResponse {
 public:
     SuccessResponse() = default;
@@ -22,6 +30,22 @@ public:
 
 protected:
     nlohmann::json response_;
+};
+
+class OpenAIResponse: public SuccessResponse {
+public:
+    OpenAIResponse() :
+        SuccessResponse() {}
+    OpenAIResponse(const std::string &response);
+
+protected:
+    std::string get_text_from_response_();
+};
+
+struct OllamaResponse: public SuccessResponse {
+    OllamaResponse() :
+        SuccessResponse() {} // needed for variants
+    OllamaResponse(const std::string &response);
 };
 
 } // namespace queries
